Add initial workings of V5 system protocol
- Added new DLC: abstraction!
  - VEX Device communication stems from a base class called VEXDevice,
  which the V5 and Cortex will inherit from. It's responsible for the
  basic packet communication the (c936/aa55 protocol)
  - Added all almost all the commands specified in the protocol, except
  for device status (21) and the key/value commands (2e/2f). Both should
  be straightforward to do, but just haven't needed to do it yet.
  - Error handling is all taken care of at the base layers of the the
  tx/rx of packet, so the actual protocol description is fairly
  straightforward - except for (21) which is why it hasn't been
  implemented. But I started on it. Will probably finish it after this
  commit.
- Having an issue uploading a generated .ini file supposedly due to a
CRC packet issue. But I know CRC handling is just fine since I've sent
out hundreds of other packets and we have positive CRC validation before
TX'ing packets. Long story short, I'm emailing James after the Christmas
holiday.
- I also want to comment that the protocol checks all bytes and does CRC
validation on the incoming data. So that's good.
- Added the beginnings of the CortexDevice to show myself that it'll be
doable. When we expose the Bootloader, we'll return an
STM32BootloaderDevice object which will have all of its commands.
Probably not going to finish fleshing it out because I don't have a
Cortex to test on.
- Copied over PROS CLI 2's terminal just because I wanted something to
play around with. To be removed later.