Replace retries, add Message class
- I wanted to remove the ugly retries parameter passing and get a nicer
solution. I created a function decorator in pros.common.utils which
wraps a function with the necessary try/catch/retry logic. We'll
decorate all of our "retry" functions with the retry decorator and it'll
just work. So all of the retry nesting got removed and replaced with the
function decorator
- I also wanted to have a cleaner way of passing all of the Tx/Rx'd data
around. I wanted to keep with the current mantra of passing back only
the relevant data (i.e. _txrx_ext_packet should only return the payload
data w/the header verified), but when we throw an exception, we want to
report all that data for debugging purposes. I created a new class
"Message" which inherits from the built-in bytes class. It allows me to
tack on the raw Tx/Rx data once at the lowest level and let it propogate
back out.
- Let PyCharm reformat all the committed files