Add new ports abstraction, serial share, terminal, COBS support
This is a big one

- Added logging into the CLI (as opposed to just hacking it in test
files). Use --log=LEVEL to set the logging verbosity level and
--log-file file LEVEL to create a log file
- Added a new port abstraction which should allow most utilities to not
care whether they're communicating with a serial share port, a straight
serial port, etc.
- Added the serial share implementation. Tested on my machine with a lot
more testing a validation to go. I'm going to create a documentation
commit soon which better describes serial share and the abstraction
layers.