Use natural (human) sorting for COM ports
Summary:
A standard ASCII-betical sort was insufficient.

Bug report by: Steve Hassenplug.

He noticed that CLI picked COM10 as the system port and COM9 as the
user port, when it should be the other way around. The defined behavior
is higher port number is the user port. The bug occurred because "COM10"
< "COM9" in ASCII-betical ordering. This commit makes sorting based on
natural human sorting.

Test Plan:
[X] Verified that COM10 was higher than COM9 and port selection works as intended.
[ ] Test on Steve's computer

Reviewers: O3 The Q Continuum, brookea

Reviewed By: O3 The Q Continuum, brookea

Subscribers: brookea

Tags: #zorp

Differential Revision: http://phabricator.purduesigbots.com/D48