Added beginning of framework for the JINX server
the server is meant to pass messages through from the connected microcontroller to our own FreeBoard-based frontend client after some filtering.

Note that this commit adds websockets (a python package) to requirements.txt, so it will be necessary to reinstall the requirements after this commit.

currently running into issues where 1) the server crashes after receiving some malformed message and 2) the server is no longer properly passing messages to the client.

I'm pretty sure that (2) will be solved once (1) is, though