🚸 Force all ./bin files to go to ./firmware in conductor create-template
Summary:
In D93, a basic library is created by compiling a libbest.a file and running `pros conductor create-template`. We need to provide the file path to the archive file, which is currently in the ./bin directory. libbest.a should really be installed into the ./firmware directory since that's where libraries are expected to be.

There were a couple of design alternatives that I could think of:
- Move ./bin/libbest.a temporarily to ./firmware/libbest.a, but we'd have to guarantee that the file is removed after creating the template, otherwise users would be linking to their own library from the archive and from source (linking errors!). This felt too risky to get right, plus isn't terribly robust/flexible
- Add additional syntax/parsing when specifying system files to the CLI that allows you to move the file from one directory to another. This would be ideal since it's the most flexible, but I couldn't think of any good way to describe this in the arguments, and this would also require a lot of code change to the create_template function
- Prevent any files from getting installed to the ./bin directory by rewriting all ./bin/* paths to ./firmware/* paths. This seems to be the simplest solution for now

Test Plan: [X] Tested building okapilib with the D93 Makefiles and okapilib.a is added correctly

Reviewers: O3 The Q Continuum, brookea

Reviewed By: O3 The Q Continuum, brookea

Subscribers: brookea

Tags: #zorp

Differential Revision: http://phabricator.purduesigbots.com/D95