Early support for CLI upgrades
Summary: - This adds the basic backbone for how the upgrade system will work. This changeset includes how the CLI will look for updates. There is not currently any logic for actually running the upgrade process - just for detection that there is an upgrade. That will come in a later diff.

Test Plan: [X] Run prosv5 upgrade and see that there's an update available (or not) using a temporary ~berman5/.www endpoint

Reviewers: O3 The Q Continuum, brookea

Reviewed By: O3 The Q Continuum, brookea

Subscribers: brookea

Tags: #zorp

Maniphest Tasks: T538

Differential Revision: http://phabricator.purduesigbots.com/D118