🍎 add binary location to path when running intercept-build on macOS
Summary: Because of the sandboxed way in which Homebrew installs python applications, prosv5 couldn't find intercept-cc and intercept-c++ when running `prosv5 build-compile-commands`. To fix this, I've added a step that will prepend the executable directory to `environment['PATH']` before attempting to invoke any of the intercept commands. This is quite similar to the way in which we were already checking and fixing Windows path issues for make.

Test Plan:
- [X] run `prosv5 build-compile-commands clean all` successfully when installed using pip/whl
- [X] run `prosv5 build-compile-commands clean all` successfully when installed using Homebrew (tested using github/develop and `--HEAD`)

Reviewers: O3 The Q Continuum, berman5

Reviewed By: O3 The Q Continuum, berman5

Tags: #zorp

Differential Revision: https://phabricator.purduesigbots.com/D208