🐛 Adjust payload_length check
Summary:
When `payload_length == 0x80` the produced bytes cause the brain to read the length as a continuation byte instead of a length byte because of the use of `>` instead of `>=`.

## Motivation:
Packets of length exactly 128 will be parsed with an invalid payload length causing the error
`ERROR - pros.cli.upload:upload - Couldn't find the response header in the device response. Got  but was expecting aa55.`

Closes purduesigbots/pros-cli3#29

Test Plan: - [x] Send a packet with a payload of 128 bytes (124 + crc) and observe a successful response from the brain

Reviewers: O3 The Q Continuum, brookea

Reviewed By: O3 The Q Continuum, brookea

Subscribers: brookea

Tags: #zorp

Differential Revision: https://phabricator.purduesigbots.com/D272