Revert changes in #38 (#55)
#### Summary:
I broke the terminal with these changes in #38.

#### Motivation:
It don't work for Jonathan. I didn't observe this when testing with my install. Not sure why it's broken on Linux but not Windows. Anywho, I just reverted the changes for both platforms since it wasn't broken to begin with.

##### References (optional):
N/A

#### Test Plan:
- [X] It work for Jonathan (on Linux)
- [X] It work for me (on Windows)
- [x] It work for someone else on Windows