Fix issues with --name
We have to create another argument to store it. I don't know why it would act the way it did -- it tried to name both the .ini file and the .bin file that, but works fine when empty. Best not to touch it and just set it back to None after copying it to the other argument.