Fix short_help descriptions
In order to have a description appear beside a list of commands (e.g. what appears when you put ``pros``, ``pros c`` or ``pros v5``) must have short_help variable in @cmd.group or @cmd.command. This is different from the help variable which is shown with groups alone (c or v5) or with --help for commands.